A wrapper function that automatically creates generalized and/or canonical shapes for TS modeling.

A wrapper function that automatically creates generalized and/or canonical shapes for TS modeling.

Usage

create_shape(shape, ...)

Arguments

shape

Shape. Details for shape specification are provided under 'Details', including mandatory arguments.

...

Additional input arguments for subsequent shape generation functions.

Value

A Shape object.

Details

The shape argument specifies what shape for the function to generate the desired shape for TS modeling. Options currently include:

Object shape	shape = ...		Parameters	Root function

Discrete/tapered cylinder	"cylinder"		length, radius	cylinder(...)

Polynomial cylinder	"polynomial_cylinder"		length, radius, polynomial	polynomial_cylinder(...)

Oblate spheroid	"oblate_spheroid"		length, radius	oblate_spheroid(...)

Prolate spheroid	"prolate_spheroid"		length, radius	prolate_spheroid(...)

Sphere	"sphere"		radius	sphere(...)

Model Parameter Definitions

length: the x-axis length of the shape.

radius: the radius of the shape when applicable.

length_radius_ratio: the length-to-radius ratio (L/A), which specifically refers to the radius at the mid-point of the cylinder and should be the maximum value. A typical L/A ratio in the literature is 16 for krill.

taper: the taper order (n), which parameterizes the tapering function reported by Chu et al. (1993) to create a tapered cylinder. The tapering order will converge on a prolate and oblate spheroid when L > 2a and L < 2a, respectively, and n = 2. A typical taper order in the literature is 10.

polynomial: the vector of arbitrary polynomial coefficients to generate a deformed cylinder as reported by Smith et al. (2013). Although listed as a mandatory argument for the polynomial cylinder function, it has a default setting that uses the sixth-degree polynomial coefficients reported by Smith et al. (2013).

Examples

create_shape("sphere", radius_body = 0.01)
#> An object of class "Sphere"
#> Slot "position_matrix":
#>             x y z          zU           zL
#>   [1,] 0.0200 0 0 0.000000000  0.000000000
#>   [2,] 0.0198 0 0 0.001989975 -0.001989975
#>   [3,] 0.0196 0 0 0.002800000 -0.002800000
#>   [4,] 0.0194 0 0 0.003411744 -0.003411744
#>   [5,] 0.0192 0 0 0.003919184 -0.003919184
#>   [6,] 0.0190 0 0 0.004358899 -0.004358899
#>   [7,] 0.0188 0 0 0.004749737 -0.004749737
#>   [8,] 0.0186 0 0 0.005102940 -0.005102940
#>   [9,] 0.0184 0 0 0.005425864 -0.005425864
#>  [10,] 0.0182 0 0 0.005723635 -0.005723635
#>  [11,] 0.0180 0 0 0.006000000 -0.006000000
#>  [12,] 0.0178 0 0 0.006257795 -0.006257795
#>  [13,] 0.0176 0 0 0.006499231 -0.006499231
#>  [14,] 0.0174 0 0 0.006726069 -0.006726069
#>  [15,] 0.0172 0 0 0.006939741 -0.006939741
#>  [16,] 0.0170 0 0 0.007141428 -0.007141428
#>  [17,] 0.0168 0 0 0.007332121 -0.007332121
#>  [18,] 0.0166 0 0 0.007512656 -0.007512656
#>  [19,] 0.0164 0 0 0.007683749 -0.007683749
#>  [20,] 0.0162 0 0 0.007846018 -0.007846018
#>  [21,] 0.0160 0 0 0.008000000 -0.008000000
#>  [22,] 0.0158 0 0 0.008146165 -0.008146165
#>  [23,] 0.0156 0 0 0.008284926 -0.008284926
#>  [24,] 0.0154 0 0 0.008416650 -0.008416650
#>  [25,] 0.0152 0 0 0.008541663 -0.008541663
#>  [26,] 0.0150 0 0 0.008660254 -0.008660254
#>  [27,] 0.0148 0 0 0.008772685 -0.008772685
#>  [28,] 0.0146 0 0 0.008879189 -0.008879189
#>  [29,] 0.0144 0 0 0.008979978 -0.008979978
#>  [30,] 0.0142 0 0 0.009075241 -0.009075241
#>  [31,] 0.0140 0 0 0.009165151 -0.009165151
#>  [32,] 0.0138 0 0 0.009249865 -0.009249865
#>  [33,] 0.0136 0 0 0.009329523 -0.009329523
#>  [34,] 0.0134 0 0 0.009404254 -0.009404254
#>  [35,] 0.0132 0 0 0.009474175 -0.009474175
#>  [36,] 0.0130 0 0 0.009539392 -0.009539392
#>  [37,] 0.0128 0 0 0.009600000 -0.009600000
#>  [38,] 0.0126 0 0 0.009656086 -0.009656086
#>  [39,] 0.0124 0 0 0.009707729 -0.009707729
#>  [40,] 0.0122 0 0 0.009754999 -0.009754999
#>  [41,] 0.0120 0 0 0.009797959 -0.009797959
#>  [42,] 0.0118 0 0 0.009836666 -0.009836666
#>  [43,] 0.0116 0 0 0.009871170 -0.009871170
#>  [44,] 0.0114 0 0 0.009901515 -0.009901515
#>  [45,] 0.0112 0 0 0.009927739 -0.009927739
#>  [46,] 0.0110 0 0 0.009949874 -0.009949874
#>  [47,] 0.0108 0 0 0.009967949 -0.009967949
#>  [48,] 0.0106 0 0 0.009981984 -0.009981984
#>  [49,] 0.0104 0 0 0.009991997 -0.009991997
#>  [50,] 0.0102 0 0 0.009998000 -0.009998000
#>  [51,] 0.0100 0 0 0.010000000 -0.010000000
#>  [52,] 0.0098 0 0 0.009998000 -0.009998000
#>  [53,] 0.0096 0 0 0.009991997 -0.009991997
#>  [54,] 0.0094 0 0 0.009981984 -0.009981984
#>  [55,] 0.0092 0 0 0.009967949 -0.009967949
#>  [56,] 0.0090 0 0 0.009949874 -0.009949874
#>  [57,] 0.0088 0 0 0.009927739 -0.009927739
#>  [58,] 0.0086 0 0 0.009901515 -0.009901515
#>  [59,] 0.0084 0 0 0.009871170 -0.009871170
#>  [60,] 0.0082 0 0 0.009836666 -0.009836666
#>  [61,] 0.0080 0 0 0.009797959 -0.009797959
#>  [62,] 0.0078 0 0 0.009754999 -0.009754999
#>  [63,] 0.0076 0 0 0.009707729 -0.009707729
#>  [64,] 0.0074 0 0 0.009656086 -0.009656086
#>  [65,] 0.0072 0 0 0.009600000 -0.009600000
#>  [66,] 0.0070 0 0 0.009539392 -0.009539392
#>  [67,] 0.0068 0 0 0.009474175 -0.009474175
#>  [68,] 0.0066 0 0 0.009404254 -0.009404254
#>  [69,] 0.0064 0 0 0.009329523 -0.009329523
#>  [70,] 0.0062 0 0 0.009249865 -0.009249865
#>  [71,] 0.0060 0 0 0.009165151 -0.009165151
#>  [72,] 0.0058 0 0 0.009075241 -0.009075241
#>  [73,] 0.0056 0 0 0.008979978 -0.008979978
#>  [74,] 0.0054 0 0 0.008879189 -0.008879189
#>  [75,] 0.0052 0 0 0.008772685 -0.008772685
#>  [76,] 0.0050 0 0 0.008660254 -0.008660254
#>  [77,] 0.0048 0 0 0.008541663 -0.008541663
#>  [78,] 0.0046 0 0 0.008416650 -0.008416650
#>  [79,] 0.0044 0 0 0.008284926 -0.008284926
#>  [80,] 0.0042 0 0 0.008146165 -0.008146165
#>  [81,] 0.0040 0 0 0.008000000 -0.008000000
#>  [82,] 0.0038 0 0 0.007846018 -0.007846018
#>  [83,] 0.0036 0 0 0.007683749 -0.007683749
#>  [84,] 0.0034 0 0 0.007512656 -0.007512656
#>  [85,] 0.0032 0 0 0.007332121 -0.007332121
#>  [86,] 0.0030 0 0 0.007141428 -0.007141428
#>  [87,] 0.0028 0 0 0.006939741 -0.006939741
#>  [88,] 0.0026 0 0 0.006726069 -0.006726069
#>  [89,] 0.0024 0 0 0.006499231 -0.006499231
#>  [90,] 0.0022 0 0 0.006257795 -0.006257795
#>  [91,] 0.0020 0 0 0.006000000 -0.006000000
#>  [92,] 0.0018 0 0 0.005723635 -0.005723635
#>  [93,] 0.0016 0 0 0.005425864 -0.005425864
#>  [94,] 0.0014 0 0 0.005102940 -0.005102940
#>  [95,] 0.0012 0 0 0.004749737 -0.004749737
#>  [96,] 0.0010 0 0 0.004358899 -0.004358899
#>  [97,] 0.0008 0 0 0.003919184 -0.003919184
#>  [98,] 0.0006 0 0 0.003411744 -0.003411744
#>  [99,] 0.0004 0 0 0.002800000 -0.002800000
#> [100,] 0.0002 0 0 0.001989975 -0.001989975
#> [101,] 0.0000 0 0 0.000000000  0.000000000
#> 
#> Slot "shape_parameters":
#> $radius_shape
#>   [1] 0.000000000 0.001989975 0.002800000 0.003411744 0.003919184 0.004358899
#>   [7] 0.004749737 0.005102940 0.005425864 0.005723635 0.006000000 0.006257795
#>  [13] 0.006499231 0.006726069 0.006939741 0.007141428 0.007332121 0.007512656
#>  [19] 0.007683749 0.007846018 0.008000000 0.008146165 0.008284926 0.008416650
#>  [25] 0.008541663 0.008660254 0.008772685 0.008879189 0.008979978 0.009075241
#>  [31] 0.009165151 0.009249865 0.009329523 0.009404254 0.009474175 0.009539392
#>  [37] 0.009600000 0.009656086 0.009707729 0.009754999 0.009797959 0.009836666
#>  [43] 0.009871170 0.009901515 0.009927739 0.009949874 0.009967949 0.009981984
#>  [49] 0.009991997 0.009998000 0.010000000 0.009998000 0.009991997 0.009981984
#>  [55] 0.009967949 0.009949874 0.009927739 0.009901515 0.009871170 0.009836666
#>  [61] 0.009797959 0.009754999 0.009707729 0.009656086 0.009600000 0.009539392
#>  [67] 0.009474175 0.009404254 0.009329523 0.009249865 0.009165151 0.009075241
#>  [73] 0.008979978 0.008879189 0.008772685 0.008660254 0.008541663 0.008416650
#>  [79] 0.008284926 0.008146165 0.008000000 0.007846018 0.007683749 0.007512656
#>  [85] 0.007332121 0.007141428 0.006939741 0.006726069 0.006499231 0.006257795
#>  [91] 0.006000000 0.005723635 0.005425864 0.005102940 0.004749737 0.004358899
#>  [97] 0.003919184 0.003411744 0.002800000 0.001989975 0.000000000
#> 
#> $diameter_shape
#> [1] 0.02
#> 
#> $radius
#> [1] 0.01
#> 
#> $n_segments
#> [1] 100
#> 
#> $diameter_units
#> [1] "m"
#> 
#> 
create_shape(
  "prolate_spheroid",
  length_body = 0.04, radius_body = 0.004
)
#> An object of class "ProlateSpheroid"
#> Slot "position_matrix":
#>             x y z           zU            zL
#>   [1,] 0.0400 0 0 0.0000000000  0.0000000000
#>   [2,] 0.0396 0 0 0.0007959899 -0.0007959899
#>   [3,] 0.0392 0 0 0.0011200000 -0.0011200000
#>   [4,] 0.0388 0 0 0.0013646978 -0.0013646978
#>   [5,] 0.0384 0 0 0.0015676734 -0.0015676734
#>   [6,] 0.0380 0 0 0.0017435596 -0.0017435596
#>   [7,] 0.0376 0 0 0.0018998947 -0.0018998947
#>   [8,] 0.0372 0 0 0.0020411761 -0.0020411761
#>   [9,] 0.0368 0 0 0.0021703456 -0.0021703456
#>  [10,] 0.0364 0 0 0.0022894541 -0.0022894541
#>  [11,] 0.0360 0 0 0.0024000000 -0.0024000000
#>  [12,] 0.0356 0 0 0.0025031181 -0.0025031181
#>  [13,] 0.0352 0 0 0.0025996923 -0.0025996923
#>  [14,] 0.0348 0 0 0.0026904275 -0.0026904275
#>  [15,] 0.0344 0 0 0.0027758963 -0.0027758963
#>  [16,] 0.0340 0 0 0.0028565714 -0.0028565714
#>  [17,] 0.0336 0 0 0.0029328484 -0.0029328484
#>  [18,] 0.0332 0 0 0.0030050624 -0.0030050624
#>  [19,] 0.0328 0 0 0.0030734996 -0.0030734996
#>  [20,] 0.0324 0 0 0.0031384072 -0.0031384072
#>  [21,] 0.0320 0 0 0.0032000000 -0.0032000000
#>  [22,] 0.0316 0 0 0.0032584659 -0.0032584659
#>  [23,] 0.0312 0 0 0.0033139704 -0.0033139704
#>  [24,] 0.0308 0 0 0.0033666601 -0.0033666601
#>  [25,] 0.0304 0 0 0.0034166650 -0.0034166650
#>  [26,] 0.0300 0 0 0.0034641016 -0.0034641016
#>  [27,] 0.0296 0 0 0.0035090740 -0.0035090740
#>  [28,] 0.0292 0 0 0.0035516757 -0.0035516757
#>  [29,] 0.0288 0 0 0.0035919911 -0.0035919911
#>  [30,] 0.0284 0 0 0.0036300964 -0.0036300964
#>  [31,] 0.0280 0 0 0.0036660606 -0.0036660606
#>  [32,] 0.0276 0 0 0.0036999459 -0.0036999459
#>  [33,] 0.0272 0 0 0.0037318092 -0.0037318092
#>  [34,] 0.0268 0 0 0.0037617017 -0.0037617017
#>  [35,] 0.0264 0 0 0.0037896702 -0.0037896702
#>  [36,] 0.0260 0 0 0.0038157568 -0.0038157568
#>  [37,] 0.0256 0 0 0.0038400000 -0.0038400000
#>  [38,] 0.0252 0 0 0.0038624345 -0.0038624345
#>  [39,] 0.0248 0 0 0.0038830916 -0.0038830916
#>  [40,] 0.0244 0 0 0.0039019995 -0.0039019995
#>  [41,] 0.0240 0 0 0.0039191836 -0.0039191836
#>  [42,] 0.0236 0 0 0.0039346664 -0.0039346664
#>  [43,] 0.0232 0 0 0.0039484681 -0.0039484681
#>  [44,] 0.0228 0 0 0.0039606060 -0.0039606060
#>  [45,] 0.0224 0 0 0.0039710956 -0.0039710956
#>  [46,] 0.0220 0 0 0.0039799497 -0.0039799497
#>  [47,] 0.0216 0 0 0.0039871795 -0.0039871795
#>  [48,] 0.0212 0 0 0.0039927935 -0.0039927935
#>  [49,] 0.0208 0 0 0.0039967987 -0.0039967987
#>  [50,] 0.0204 0 0 0.0039991999 -0.0039991999
#>  [51,] 0.0200 0 0 0.0040000000 -0.0040000000
#>  [52,] 0.0196 0 0 0.0039991999 -0.0039991999
#>  [53,] 0.0192 0 0 0.0039967987 -0.0039967987
#>  [54,] 0.0188 0 0 0.0039927935 -0.0039927935
#>  [55,] 0.0184 0 0 0.0039871795 -0.0039871795
#>  [56,] 0.0180 0 0 0.0039799497 -0.0039799497
#>  [57,] 0.0176 0 0 0.0039710956 -0.0039710956
#>  [58,] 0.0172 0 0 0.0039606060 -0.0039606060
#>  [59,] 0.0168 0 0 0.0039484681 -0.0039484681
#>  [60,] 0.0164 0 0 0.0039346664 -0.0039346664
#>  [61,] 0.0160 0 0 0.0039191836 -0.0039191836
#>  [62,] 0.0156 0 0 0.0039019995 -0.0039019995
#>  [63,] 0.0152 0 0 0.0038830916 -0.0038830916
#>  [64,] 0.0148 0 0 0.0038624345 -0.0038624345
#>  [65,] 0.0144 0 0 0.0038400000 -0.0038400000
#>  [66,] 0.0140 0 0 0.0038157568 -0.0038157568
#>  [67,] 0.0136 0 0 0.0037896702 -0.0037896702
#>  [68,] 0.0132 0 0 0.0037617017 -0.0037617017
#>  [69,] 0.0128 0 0 0.0037318092 -0.0037318092
#>  [70,] 0.0124 0 0 0.0036999459 -0.0036999459
#>  [71,] 0.0120 0 0 0.0036660606 -0.0036660606
#>  [72,] 0.0116 0 0 0.0036300964 -0.0036300964
#>  [73,] 0.0112 0 0 0.0035919911 -0.0035919911
#>  [74,] 0.0108 0 0 0.0035516757 -0.0035516757
#>  [75,] 0.0104 0 0 0.0035090740 -0.0035090740
#>  [76,] 0.0100 0 0 0.0034641016 -0.0034641016
#>  [77,] 0.0096 0 0 0.0034166650 -0.0034166650
#>  [78,] 0.0092 0 0 0.0033666601 -0.0033666601
#>  [79,] 0.0088 0 0 0.0033139704 -0.0033139704
#>  [80,] 0.0084 0 0 0.0032584659 -0.0032584659
#>  [81,] 0.0080 0 0 0.0032000000 -0.0032000000
#>  [82,] 0.0076 0 0 0.0031384072 -0.0031384072
#>  [83,] 0.0072 0 0 0.0030734996 -0.0030734996
#>  [84,] 0.0068 0 0 0.0030050624 -0.0030050624
#>  [85,] 0.0064 0 0 0.0029328484 -0.0029328484
#>  [86,] 0.0060 0 0 0.0028565714 -0.0028565714
#>  [87,] 0.0056 0 0 0.0027758963 -0.0027758963
#>  [88,] 0.0052 0 0 0.0026904275 -0.0026904275
#>  [89,] 0.0048 0 0 0.0025996923 -0.0025996923
#>  [90,] 0.0044 0 0 0.0025031181 -0.0025031181
#>  [91,] 0.0040 0 0 0.0024000000 -0.0024000000
#>  [92,] 0.0036 0 0 0.0022894541 -0.0022894541
#>  [93,] 0.0032 0 0 0.0021703456 -0.0021703456
#>  [94,] 0.0028 0 0 0.0020411761 -0.0020411761
#>  [95,] 0.0024 0 0 0.0018998947 -0.0018998947
#>  [96,] 0.0020 0 0 0.0017435596 -0.0017435596
#>  [97,] 0.0016 0 0 0.0015676734 -0.0015676734
#>  [98,] 0.0012 0 0 0.0013646978 -0.0013646978
#>  [99,] 0.0008 0 0 0.0011200000 -0.0011200000
#> [100,] 0.0004 0 0 0.0007959899 -0.0007959899
#> [101,] 0.0000 0 0 0.0000000000  0.0000000000
#> 
#> Slot "shape_parameters":
#> $length
#> [1] 0.04
#> 
#> $radius
#>   [1] 0.0000000000 0.0007959899 0.0011200000 0.0013646978 0.0015676734
#>   [6] 0.0017435596 0.0018998947 0.0020411761 0.0021703456 0.0022894541
#>  [11] 0.0024000000 0.0025031181 0.0025996923 0.0026904275 0.0027758963
#>  [16] 0.0028565714 0.0029328484 0.0030050624 0.0030734996 0.0031384072
#>  [21] 0.0032000000 0.0032584659 0.0033139704 0.0033666601 0.0034166650
#>  [26] 0.0034641016 0.0035090740 0.0035516757 0.0035919911 0.0036300964
#>  [31] 0.0036660606 0.0036999459 0.0037318092 0.0037617017 0.0037896702
#>  [36] 0.0038157568 0.0038400000 0.0038624345 0.0038830916 0.0039019995
#>  [41] 0.0039191836 0.0039346664 0.0039484681 0.0039606060 0.0039710956
#>  [46] 0.0039799497 0.0039871795 0.0039927935 0.0039967987 0.0039991999
#>  [51] 0.0040000000 0.0039991999 0.0039967987 0.0039927935 0.0039871795
#>  [56] 0.0039799497 0.0039710956 0.0039606060 0.0039484681 0.0039346664
#>  [61] 0.0039191836 0.0039019995 0.0038830916 0.0038624345 0.0038400000
#>  [66] 0.0038157568 0.0037896702 0.0037617017 0.0037318092 0.0036999459
#>  [71] 0.0036660606 0.0036300964 0.0035919911 0.0035516757 0.0035090740
#>  [76] 0.0034641016 0.0034166650 0.0033666601 0.0033139704 0.0032584659
#>  [81] 0.0032000000 0.0031384072 0.0030734996 0.0030050624 0.0029328484
#>  [86] 0.0028565714 0.0027758963 0.0026904275 0.0025996923 0.0025031181
#>  [91] 0.0024000000 0.0022894541 0.0021703456 0.0020411761 0.0018998947
#>  [96] 0.0017435596 0.0015676734 0.0013646978 0.0011200000 0.0007959899
#> [101] 0.0000000000
#> 
#> $semimajor_length
#> [1] 0.02
#> 
#> $semiminor_length
#> [1] 0.004
#> 
#> $length_radius_ratio
#> [1] 10
#> 
#> $n_segments
#> [1] 100
#> 
#> $length_units
#> [1] "m"
#> 
#> 
create_shape(
  "oblate_spheroid",
  length_body = 0.012, radius_body = 0.01
)
#> An object of class "OblateSpheroid"
#> Slot "position_matrix":
#>              x y z          zU           zL
#>   [1,] 0.01200 0 0 0.000000000  0.000000000
#>   [2,] 0.01188 0 0 0.001989975 -0.001989975
#>   [3,] 0.01176 0 0 0.002800000 -0.002800000
#>   [4,] 0.01164 0 0 0.003411744 -0.003411744
#>   [5,] 0.01152 0 0 0.003919184 -0.003919184
#>   [6,] 0.01140 0 0 0.004358899 -0.004358899
#>   [7,] 0.01128 0 0 0.004749737 -0.004749737
#>   [8,] 0.01116 0 0 0.005102940 -0.005102940
#>   [9,] 0.01104 0 0 0.005425864 -0.005425864
#>  [10,] 0.01092 0 0 0.005723635 -0.005723635
#>  [11,] 0.01080 0 0 0.006000000 -0.006000000
#>  [12,] 0.01068 0 0 0.006257795 -0.006257795
#>  [13,] 0.01056 0 0 0.006499231 -0.006499231
#>  [14,] 0.01044 0 0 0.006726069 -0.006726069
#>  [15,] 0.01032 0 0 0.006939741 -0.006939741
#>  [16,] 0.01020 0 0 0.007141428 -0.007141428
#>  [17,] 0.01008 0 0 0.007332121 -0.007332121
#>  [18,] 0.00996 0 0 0.007512656 -0.007512656
#>  [19,] 0.00984 0 0 0.007683749 -0.007683749
#>  [20,] 0.00972 0 0 0.007846018 -0.007846018
#>  [21,] 0.00960 0 0 0.008000000 -0.008000000
#>  [22,] 0.00948 0 0 0.008146165 -0.008146165
#>  [23,] 0.00936 0 0 0.008284926 -0.008284926
#>  [24,] 0.00924 0 0 0.008416650 -0.008416650
#>  [25,] 0.00912 0 0 0.008541663 -0.008541663
#>  [26,] 0.00900 0 0 0.008660254 -0.008660254
#>  [27,] 0.00888 0 0 0.008772685 -0.008772685
#>  [28,] 0.00876 0 0 0.008879189 -0.008879189
#>  [29,] 0.00864 0 0 0.008979978 -0.008979978
#>  [30,] 0.00852 0 0 0.009075241 -0.009075241
#>  [31,] 0.00840 0 0 0.009165151 -0.009165151
#>  [32,] 0.00828 0 0 0.009249865 -0.009249865
#>  [33,] 0.00816 0 0 0.009329523 -0.009329523
#>  [34,] 0.00804 0 0 0.009404254 -0.009404254
#>  [35,] 0.00792 0 0 0.009474175 -0.009474175
#>  [36,] 0.00780 0 0 0.009539392 -0.009539392
#>  [37,] 0.00768 0 0 0.009600000 -0.009600000
#>  [38,] 0.00756 0 0 0.009656086 -0.009656086
#>  [39,] 0.00744 0 0 0.009707729 -0.009707729
#>  [40,] 0.00732 0 0 0.009754999 -0.009754999
#>  [41,] 0.00720 0 0 0.009797959 -0.009797959
#>  [42,] 0.00708 0 0 0.009836666 -0.009836666
#>  [43,] 0.00696 0 0 0.009871170 -0.009871170
#>  [44,] 0.00684 0 0 0.009901515 -0.009901515
#>  [45,] 0.00672 0 0 0.009927739 -0.009927739
#>  [46,] 0.00660 0 0 0.009949874 -0.009949874
#>  [47,] 0.00648 0 0 0.009967949 -0.009967949
#>  [48,] 0.00636 0 0 0.009981984 -0.009981984
#>  [49,] 0.00624 0 0 0.009991997 -0.009991997
#>  [50,] 0.00612 0 0 0.009998000 -0.009998000
#>  [51,] 0.00600 0 0 0.010000000 -0.010000000
#>  [52,] 0.00588 0 0 0.009998000 -0.009998000
#>  [53,] 0.00576 0 0 0.009991997 -0.009991997
#>  [54,] 0.00564 0 0 0.009981984 -0.009981984
#>  [55,] 0.00552 0 0 0.009967949 -0.009967949
#>  [56,] 0.00540 0 0 0.009949874 -0.009949874
#>  [57,] 0.00528 0 0 0.009927739 -0.009927739
#>  [58,] 0.00516 0 0 0.009901515 -0.009901515
#>  [59,] 0.00504 0 0 0.009871170 -0.009871170
#>  [60,] 0.00492 0 0 0.009836666 -0.009836666
#>  [61,] 0.00480 0 0 0.009797959 -0.009797959
#>  [62,] 0.00468 0 0 0.009754999 -0.009754999
#>  [63,] 0.00456 0 0 0.009707729 -0.009707729
#>  [64,] 0.00444 0 0 0.009656086 -0.009656086
#>  [65,] 0.00432 0 0 0.009600000 -0.009600000
#>  [66,] 0.00420 0 0 0.009539392 -0.009539392
#>  [67,] 0.00408 0 0 0.009474175 -0.009474175
#>  [68,] 0.00396 0 0 0.009404254 -0.009404254
#>  [69,] 0.00384 0 0 0.009329523 -0.009329523
#>  [70,] 0.00372 0 0 0.009249865 -0.009249865
#>  [71,] 0.00360 0 0 0.009165151 -0.009165151
#>  [72,] 0.00348 0 0 0.009075241 -0.009075241
#>  [73,] 0.00336 0 0 0.008979978 -0.008979978
#>  [74,] 0.00324 0 0 0.008879189 -0.008879189
#>  [75,] 0.00312 0 0 0.008772685 -0.008772685
#>  [76,] 0.00300 0 0 0.008660254 -0.008660254
#>  [77,] 0.00288 0 0 0.008541663 -0.008541663
#>  [78,] 0.00276 0 0 0.008416650 -0.008416650
#>  [79,] 0.00264 0 0 0.008284926 -0.008284926
#>  [80,] 0.00252 0 0 0.008146165 -0.008146165
#>  [81,] 0.00240 0 0 0.008000000 -0.008000000
#>  [82,] 0.00228 0 0 0.007846018 -0.007846018
#>  [83,] 0.00216 0 0 0.007683749 -0.007683749
#>  [84,] 0.00204 0 0 0.007512656 -0.007512656
#>  [85,] 0.00192 0 0 0.007332121 -0.007332121
#>  [86,] 0.00180 0 0 0.007141428 -0.007141428
#>  [87,] 0.00168 0 0 0.006939741 -0.006939741
#>  [88,] 0.00156 0 0 0.006726069 -0.006726069
#>  [89,] 0.00144 0 0 0.006499231 -0.006499231
#>  [90,] 0.00132 0 0 0.006257795 -0.006257795
#>  [91,] 0.00120 0 0 0.006000000 -0.006000000
#>  [92,] 0.00108 0 0 0.005723635 -0.005723635
#>  [93,] 0.00096 0 0 0.005425864 -0.005425864
#>  [94,] 0.00084 0 0 0.005102940 -0.005102940
#>  [95,] 0.00072 0 0 0.004749737 -0.004749737
#>  [96,] 0.00060 0 0 0.004358899 -0.004358899
#>  [97,] 0.00048 0 0 0.003919184 -0.003919184
#>  [98,] 0.00036 0 0 0.003411744 -0.003411744
#>  [99,] 0.00024 0 0 0.002800000 -0.002800000
#> [100,] 0.00012 0 0 0.001989975 -0.001989975
#> [101,] 0.00000 0 0 0.000000000  0.000000000
#> 
#> Slot "shape_parameters":
#> $length
#> [1] 0.012
#> 
#> $radius
#>   [1] 0.000000000 0.001989975 0.002800000 0.003411744 0.003919184 0.004358899
#>   [7] 0.004749737 0.005102940 0.005425864 0.005723635 0.006000000 0.006257795
#>  [13] 0.006499231 0.006726069 0.006939741 0.007141428 0.007332121 0.007512656
#>  [19] 0.007683749 0.007846018 0.008000000 0.008146165 0.008284926 0.008416650
#>  [25] 0.008541663 0.008660254 0.008772685 0.008879189 0.008979978 0.009075241
#>  [31] 0.009165151 0.009249865 0.009329523 0.009404254 0.009474175 0.009539392
#>  [37] 0.009600000 0.009656086 0.009707729 0.009754999 0.009797959 0.009836666
#>  [43] 0.009871170 0.009901515 0.009927739 0.009949874 0.009967949 0.009981984
#>  [49] 0.009991997 0.009998000 0.010000000 0.009998000 0.009991997 0.009981984
#>  [55] 0.009967949 0.009949874 0.009927739 0.009901515 0.009871170 0.009836666
#>  [61] 0.009797959 0.009754999 0.009707729 0.009656086 0.009600000 0.009539392
#>  [67] 0.009474175 0.009404254 0.009329523 0.009249865 0.009165151 0.009075241
#>  [73] 0.008979978 0.008879189 0.008772685 0.008660254 0.008541663 0.008416650
#>  [79] 0.008284926 0.008146165 0.008000000 0.007846018 0.007683749 0.007512656
#>  [85] 0.007332121 0.007141428 0.006939741 0.006726069 0.006499231 0.006257795
#>  [91] 0.006000000 0.005723635 0.005425864 0.005102940 0.004749737 0.004358899
#>  [97] 0.003919184 0.003411744 0.002800000 0.001989975 0.000000000
#> 
#> $semimajor_length
#> [1] 0.01
#> 
#> $semiminor_length
#> [1] 0.006
#> 
#> $length_radius_ratio
#> [1] 1.2
#> 
#> $n_segments
#> [1] 100
#> 
#> $length_units
#> [1] "m"
#> 
#> 
create_shape(
  "cylinder",
  length_body = 0.05, radius_body = 0.003
)
#> An object of class "Cylinder"
#> Slot "position_matrix":
#>             x y z    zU     zL
#>   [1,] 0.0500 0 0 0.003 -0.003
#>   [2,] 0.0495 0 0 0.003 -0.003
#>   [3,] 0.0490 0 0 0.003 -0.003
#>   [4,] 0.0485 0 0 0.003 -0.003
#>   [5,] 0.0480 0 0 0.003 -0.003
#>   [6,] 0.0475 0 0 0.003 -0.003
#>   [7,] 0.0470 0 0 0.003 -0.003
#>   [8,] 0.0465 0 0 0.003 -0.003
#>   [9,] 0.0460 0 0 0.003 -0.003
#>  [10,] 0.0455 0 0 0.003 -0.003
#>  [11,] 0.0450 0 0 0.003 -0.003
#>  [12,] 0.0445 0 0 0.003 -0.003
#>  [13,] 0.0440 0 0 0.003 -0.003
#>  [14,] 0.0435 0 0 0.003 -0.003
#>  [15,] 0.0430 0 0 0.003 -0.003
#>  [16,] 0.0425 0 0 0.003 -0.003
#>  [17,] 0.0420 0 0 0.003 -0.003
#>  [18,] 0.0415 0 0 0.003 -0.003
#>  [19,] 0.0410 0 0 0.003 -0.003
#>  [20,] 0.0405 0 0 0.003 -0.003
#>  [21,] 0.0400 0 0 0.003 -0.003
#>  [22,] 0.0395 0 0 0.003 -0.003
#>  [23,] 0.0390 0 0 0.003 -0.003
#>  [24,] 0.0385 0 0 0.003 -0.003
#>  [25,] 0.0380 0 0 0.003 -0.003
#>  [26,] 0.0375 0 0 0.003 -0.003
#>  [27,] 0.0370 0 0 0.003 -0.003
#>  [28,] 0.0365 0 0 0.003 -0.003
#>  [29,] 0.0360 0 0 0.003 -0.003
#>  [30,] 0.0355 0 0 0.003 -0.003
#>  [31,] 0.0350 0 0 0.003 -0.003
#>  [32,] 0.0345 0 0 0.003 -0.003
#>  [33,] 0.0340 0 0 0.003 -0.003
#>  [34,] 0.0335 0 0 0.003 -0.003
#>  [35,] 0.0330 0 0 0.003 -0.003
#>  [36,] 0.0325 0 0 0.003 -0.003
#>  [37,] 0.0320 0 0 0.003 -0.003
#>  [38,] 0.0315 0 0 0.003 -0.003
#>  [39,] 0.0310 0 0 0.003 -0.003
#>  [40,] 0.0305 0 0 0.003 -0.003
#>  [41,] 0.0300 0 0 0.003 -0.003
#>  [42,] 0.0295 0 0 0.003 -0.003
#>  [43,] 0.0290 0 0 0.003 -0.003
#>  [44,] 0.0285 0 0 0.003 -0.003
#>  [45,] 0.0280 0 0 0.003 -0.003
#>  [46,] 0.0275 0 0 0.003 -0.003
#>  [47,] 0.0270 0 0 0.003 -0.003
#>  [48,] 0.0265 0 0 0.003 -0.003
#>  [49,] 0.0260 0 0 0.003 -0.003
#>  [50,] 0.0255 0 0 0.003 -0.003
#>  [51,] 0.0250 0 0 0.003 -0.003
#>  [52,] 0.0245 0 0 0.003 -0.003
#>  [53,] 0.0240 0 0 0.003 -0.003
#>  [54,] 0.0235 0 0 0.003 -0.003
#>  [55,] 0.0230 0 0 0.003 -0.003
#>  [56,] 0.0225 0 0 0.003 -0.003
#>  [57,] 0.0220 0 0 0.003 -0.003
#>  [58,] 0.0215 0 0 0.003 -0.003
#>  [59,] 0.0210 0 0 0.003 -0.003
#>  [60,] 0.0205 0 0 0.003 -0.003
#>  [61,] 0.0200 0 0 0.003 -0.003
#>  [62,] 0.0195 0 0 0.003 -0.003
#>  [63,] 0.0190 0 0 0.003 -0.003
#>  [64,] 0.0185 0 0 0.003 -0.003
#>  [65,] 0.0180 0 0 0.003 -0.003
#>  [66,] 0.0175 0 0 0.003 -0.003
#>  [67,] 0.0170 0 0 0.003 -0.003
#>  [68,] 0.0165 0 0 0.003 -0.003
#>  [69,] 0.0160 0 0 0.003 -0.003
#>  [70,] 0.0155 0 0 0.003 -0.003
#>  [71,] 0.0150 0 0 0.003 -0.003
#>  [72,] 0.0145 0 0 0.003 -0.003
#>  [73,] 0.0140 0 0 0.003 -0.003
#>  [74,] 0.0135 0 0 0.003 -0.003
#>  [75,] 0.0130 0 0 0.003 -0.003
#>  [76,] 0.0125 0 0 0.003 -0.003
#>  [77,] 0.0120 0 0 0.003 -0.003
#>  [78,] 0.0115 0 0 0.003 -0.003
#>  [79,] 0.0110 0 0 0.003 -0.003
#>  [80,] 0.0105 0 0 0.003 -0.003
#>  [81,] 0.0100 0 0 0.003 -0.003
#>  [82,] 0.0095 0 0 0.003 -0.003
#>  [83,] 0.0090 0 0 0.003 -0.003
#>  [84,] 0.0085 0 0 0.003 -0.003
#>  [85,] 0.0080 0 0 0.003 -0.003
#>  [86,] 0.0075 0 0 0.003 -0.003
#>  [87,] 0.0070 0 0 0.003 -0.003
#>  [88,] 0.0065 0 0 0.003 -0.003
#>  [89,] 0.0060 0 0 0.003 -0.003
#>  [90,] 0.0055 0 0 0.003 -0.003
#>  [91,] 0.0050 0 0 0.003 -0.003
#>  [92,] 0.0045 0 0 0.003 -0.003
#>  [93,] 0.0040 0 0 0.003 -0.003
#>  [94,] 0.0035 0 0 0.003 -0.003
#>  [95,] 0.0030 0 0 0.003 -0.003
#>  [96,] 0.0025 0 0 0.003 -0.003
#>  [97,] 0.0020 0 0 0.003 -0.003
#>  [98,] 0.0015 0 0 0.003 -0.003
#>  [99,] 0.0010 0 0 0.003 -0.003
#> [100,] 0.0005 0 0 0.003 -0.003
#> [101,] 0.0000 0 0 0.003 -0.003
#> 
#> Slot "shape_parameters":
#> $length
#> [1] 0.05
#> 
#> $radius
#>   [1] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#>  [13] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#>  [25] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#>  [37] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#>  [49] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#>  [61] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#>  [73] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#>  [85] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#>  [97] 0.003 0.003 0.003 0.003 0.003
#> 
#> $length_radius_ratio
#> [1] 16.66667
#> 
#> $n_segments
#> [1] 100
#> 
#> $taper_order
#> [1] NA
#> 
#> $radius_curvature_ratio
#> [1] NA
#> 
#> $length_units
#> [1] "m"
#> 
#>