Creates a cylinder.

Creates a cylinder.

Usage

cylinder(length_body, radius_body = NULL, length_radius_ratio = NULL,
  taper = NULL, radius_curvature_ratio = NULL, n_segments = 100,
  length_units = "m")

Arguments

length_body

Length (m).

radius_body

Maximum/uniform radius (m).

length_radius_ratio

Optional ratio input when radius is not explicitly known.

taper

Optional input that is the degree of taper to round ends of the cylinder.

radius_curvature_ratio

Optional curvature ratio metadata for rounded cylinder-end workflows.

n_segments

Number of segments to discretize object shape. Defaults to 1e2 segments.

length_units

Units (default is meters, "m").

Value

Creates the position vector for a tapered or untapered cylinder.

See also

Examples

cylinder(length_body = 0.05, radius_body = 0.003, n_segments = 80)
#> An object of class "Cylinder"
#> Slot "position_matrix":
#>              x y z    zU     zL
#>  [1,] 0.050000 0 0 0.003 -0.003
#>  [2,] 0.049375 0 0 0.003 -0.003
#>  [3,] 0.048750 0 0 0.003 -0.003
#>  [4,] 0.048125 0 0 0.003 -0.003
#>  [5,] 0.047500 0 0 0.003 -0.003
#>  [6,] 0.046875 0 0 0.003 -0.003
#>  [7,] 0.046250 0 0 0.003 -0.003
#>  [8,] 0.045625 0 0 0.003 -0.003
#>  [9,] 0.045000 0 0 0.003 -0.003
#> [10,] 0.044375 0 0 0.003 -0.003
#> [11,] 0.043750 0 0 0.003 -0.003
#> [12,] 0.043125 0 0 0.003 -0.003
#> [13,] 0.042500 0 0 0.003 -0.003
#> [14,] 0.041875 0 0 0.003 -0.003
#> [15,] 0.041250 0 0 0.003 -0.003
#> [16,] 0.040625 0 0 0.003 -0.003
#> [17,] 0.040000 0 0 0.003 -0.003
#> [18,] 0.039375 0 0 0.003 -0.003
#> [19,] 0.038750 0 0 0.003 -0.003
#> [20,] 0.038125 0 0 0.003 -0.003
#> [21,] 0.037500 0 0 0.003 -0.003
#> [22,] 0.036875 0 0 0.003 -0.003
#> [23,] 0.036250 0 0 0.003 -0.003
#> [24,] 0.035625 0 0 0.003 -0.003
#> [25,] 0.035000 0 0 0.003 -0.003
#> [26,] 0.034375 0 0 0.003 -0.003
#> [27,] 0.033750 0 0 0.003 -0.003
#> [28,] 0.033125 0 0 0.003 -0.003
#> [29,] 0.032500 0 0 0.003 -0.003
#> [30,] 0.031875 0 0 0.003 -0.003
#> [31,] 0.031250 0 0 0.003 -0.003
#> [32,] 0.030625 0 0 0.003 -0.003
#> [33,] 0.030000 0 0 0.003 -0.003
#> [34,] 0.029375 0 0 0.003 -0.003
#> [35,] 0.028750 0 0 0.003 -0.003
#> [36,] 0.028125 0 0 0.003 -0.003
#> [37,] 0.027500 0 0 0.003 -0.003
#> [38,] 0.026875 0 0 0.003 -0.003
#> [39,] 0.026250 0 0 0.003 -0.003
#> [40,] 0.025625 0 0 0.003 -0.003
#> [41,] 0.025000 0 0 0.003 -0.003
#> [42,] 0.024375 0 0 0.003 -0.003
#> [43,] 0.023750 0 0 0.003 -0.003
#> [44,] 0.023125 0 0 0.003 -0.003
#> [45,] 0.022500 0 0 0.003 -0.003
#> [46,] 0.021875 0 0 0.003 -0.003
#> [47,] 0.021250 0 0 0.003 -0.003
#> [48,] 0.020625 0 0 0.003 -0.003
#> [49,] 0.020000 0 0 0.003 -0.003
#> [50,] 0.019375 0 0 0.003 -0.003
#> [51,] 0.018750 0 0 0.003 -0.003
#> [52,] 0.018125 0 0 0.003 -0.003
#> [53,] 0.017500 0 0 0.003 -0.003
#> [54,] 0.016875 0 0 0.003 -0.003
#> [55,] 0.016250 0 0 0.003 -0.003
#> [56,] 0.015625 0 0 0.003 -0.003
#> [57,] 0.015000 0 0 0.003 -0.003
#> [58,] 0.014375 0 0 0.003 -0.003
#> [59,] 0.013750 0 0 0.003 -0.003
#> [60,] 0.013125 0 0 0.003 -0.003
#> [61,] 0.012500 0 0 0.003 -0.003
#> [62,] 0.011875 0 0 0.003 -0.003
#> [63,] 0.011250 0 0 0.003 -0.003
#> [64,] 0.010625 0 0 0.003 -0.003
#> [65,] 0.010000 0 0 0.003 -0.003
#> [66,] 0.009375 0 0 0.003 -0.003
#> [67,] 0.008750 0 0 0.003 -0.003
#> [68,] 0.008125 0 0 0.003 -0.003
#> [69,] 0.007500 0 0 0.003 -0.003
#> [70,] 0.006875 0 0 0.003 -0.003
#> [71,] 0.006250 0 0 0.003 -0.003
#> [72,] 0.005625 0 0 0.003 -0.003
#> [73,] 0.005000 0 0 0.003 -0.003
#> [74,] 0.004375 0 0 0.003 -0.003
#> [75,] 0.003750 0 0 0.003 -0.003
#> [76,] 0.003125 0 0 0.003 -0.003
#> [77,] 0.002500 0 0 0.003 -0.003
#> [78,] 0.001875 0 0 0.003 -0.003
#> [79,] 0.001250 0 0 0.003 -0.003
#> [80,] 0.000625 0 0 0.003 -0.003
#> [81,] 0.000000 0 0 0.003 -0.003
#> 
#> Slot "shape_parameters":
#> $length
#> [1] 0.05
#> 
#> $radius
#>  [1] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#> [13] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#> [25] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#> [37] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#> [49] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#> [61] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#> [73] 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003 0.003
#> 
#> $length_radius_ratio
#> [1] 16.66667
#> 
#> $n_segments
#> [1] 80
#> 
#> $taper_order
#> [1] NA
#> 
#> $radius_curvature_ratio
#> [1] NA
#> 
#> $length_units
#> [1] "m"
#> 
#>