Creates an oblate spheroid.

Creates an oblate spheroid.

Usage

oblate_spheroid(
  length_body = NULL,
  radius_body = NULL,
  length_radius_ratio = NULL,
  semimajor_length = NULL,
  semiminor_length = NULL,
  n_segments = 100,
  length_units = "m"
)

Arguments

length_body

Body-axis length (m).

radius_body

Maximum equatorial radius (m).

length_radius_ratio

Optional ratio input when radius is not explicitly known.

semimajor_length

Optional alias for maximum equatorial radius (m).

semiminor_length

Optional alias for body-axis semi-length (m).

n_segments

Number of segments to discretize object shape. Defaults to 100 segments.

length_units

Units for body matrix (defaults to m).

Value

Creates the position vector for an oblate spheroid object of defined body-axis length and maximum equatorial radius.

See also

Examples

oblate_spheroid(
  length_body = 0.012, radius_body = 0.01, n_segments = 60
)
#> An object of class "OblateSpheroid"
#> Slot "position_matrix":
#>            x y z          zU           zL
#>  [1,] 0.0120 0 0 0.000000000  0.000000000
#>  [2,] 0.0118 0 0 0.002560382 -0.002560382
#>  [3,] 0.0116 0 0 0.003590110 -0.003590110
#>  [4,] 0.0114 0 0 0.004358899 -0.004358899
#>  [5,] 0.0112 0 0 0.004988877 -0.004988877
#>  [6,] 0.0110 0 0 0.005527708 -0.005527708
#>  [7,] 0.0108 0 0 0.006000000 -0.006000000
#>  [8,] 0.0106 0 0 0.006420453 -0.006420453
#>  [9,] 0.0104 0 0 0.006798693 -0.006798693
#> [10,] 0.0102 0 0 0.007141428 -0.007141428
#> [11,] 0.0100 0 0 0.007453560 -0.007453560
#> [12,] 0.0098 0 0 0.007738791 -0.007738791
#> [13,] 0.0096 0 0 0.008000000 -0.008000000
#> [14,] 0.0094 0 0 0.008239471 -0.008239471
#> [15,] 0.0092 0 0 0.008459052 -0.008459052
#> [16,] 0.0090 0 0 0.008660254 -0.008660254
#> [17,] 0.0088 0 0 0.008844333 -0.008844333
#> [18,] 0.0086 0 0 0.009012337 -0.009012337
#> [19,] 0.0084 0 0 0.009165151 -0.009165151
#> [20,] 0.0082 0 0 0.009303524 -0.009303524
#> [21,] 0.0080 0 0 0.009428090 -0.009428090
#> [22,] 0.0078 0 0 0.009539392 -0.009539392
#> [23,] 0.0076 0 0 0.009637888 -0.009637888
#> [24,] 0.0074 0 0 0.009723968 -0.009723968
#> [25,] 0.0072 0 0 0.009797959 -0.009797959
#> [26,] 0.0070 0 0 0.009860133 -0.009860133
#> [27,] 0.0068 0 0 0.009910712 -0.009910712
#> [28,] 0.0066 0 0 0.009949874 -0.009949874
#> [29,] 0.0064 0 0 0.009977753 -0.009977753
#> [30,] 0.0062 0 0 0.009994443 -0.009994443
#> [31,] 0.0060 0 0 0.010000000 -0.010000000
#> [32,] 0.0058 0 0 0.009994443 -0.009994443
#> [33,] 0.0056 0 0 0.009977753 -0.009977753
#> [34,] 0.0054 0 0 0.009949874 -0.009949874
#> [35,] 0.0052 0 0 0.009910712 -0.009910712
#> [36,] 0.0050 0 0 0.009860133 -0.009860133
#> [37,] 0.0048 0 0 0.009797959 -0.009797959
#> [38,] 0.0046 0 0 0.009723968 -0.009723968
#> [39,] 0.0044 0 0 0.009637888 -0.009637888
#> [40,] 0.0042 0 0 0.009539392 -0.009539392
#> [41,] 0.0040 0 0 0.009428090 -0.009428090
#> [42,] 0.0038 0 0 0.009303524 -0.009303524
#> [43,] 0.0036 0 0 0.009165151 -0.009165151
#> [44,] 0.0034 0 0 0.009012337 -0.009012337
#> [45,] 0.0032 0 0 0.008844333 -0.008844333
#> [46,] 0.0030 0 0 0.008660254 -0.008660254
#> [47,] 0.0028 0 0 0.008459052 -0.008459052
#> [48,] 0.0026 0 0 0.008239471 -0.008239471
#> [49,] 0.0024 0 0 0.008000000 -0.008000000
#> [50,] 0.0022 0 0 0.007738791 -0.007738791
#> [51,] 0.0020 0 0 0.007453560 -0.007453560
#> [52,] 0.0018 0 0 0.007141428 -0.007141428
#> [53,] 0.0016 0 0 0.006798693 -0.006798693
#> [54,] 0.0014 0 0 0.006420453 -0.006420453
#> [55,] 0.0012 0 0 0.006000000 -0.006000000
#> [56,] 0.0010 0 0 0.005527708 -0.005527708
#> [57,] 0.0008 0 0 0.004988877 -0.004988877
#> [58,] 0.0006 0 0 0.004358899 -0.004358899
#> [59,] 0.0004 0 0 0.003590110 -0.003590110
#> [60,] 0.0002 0 0 0.002560382 -0.002560382
#> [61,] 0.0000 0 0 0.000000000  0.000000000
#> 
#> Slot "shape_parameters":
#> $length
#> [1] 0.012
#> 
#> $radius
#>  [1] 0.000000000 0.002560382 0.003590110 0.004358899 0.004988877 0.005527708
#>  [7] 0.006000000 0.006420453 0.006798693 0.007141428 0.007453560 0.007738791
#> [13] 0.008000000 0.008239471 0.008459052 0.008660254 0.008844333 0.009012337
#> [19] 0.009165151 0.009303524 0.009428090 0.009539392 0.009637888 0.009723968
#> [25] 0.009797959 0.009860133 0.009910712 0.009949874 0.009977753 0.009994443
#> [31] 0.010000000 0.009994443 0.009977753 0.009949874 0.009910712 0.009860133
#> [37] 0.009797959 0.009723968 0.009637888 0.009539392 0.009428090 0.009303524
#> [43] 0.009165151 0.009012337 0.008844333 0.008660254 0.008459052 0.008239471
#> [49] 0.008000000 0.007738791 0.007453560 0.007141428 0.006798693 0.006420453
#> [55] 0.006000000 0.005527708 0.004988877 0.004358899 0.003590110 0.002560382
#> [61] 0.000000000
#> 
#> $semimajor_length
#> [1] 0.01
#> 
#> $semiminor_length
#> [1] 0.006
#> 
#> $length_radius_ratio
#> [1] 1.2
#> 
#> $n_segments
#> [1] 60
#> 
#> $length_units
#> [1] "m"
#> 
#> 
oblate_spheroid(
  semiminor_length = 0.006, semimajor_length = 0.01
)
#> An object of class "OblateSpheroid"
#> Slot "position_matrix":
#>              x y z          zU           zL
#>   [1,] 0.01200 0 0 0.000000000  0.000000000
#>   [2,] 0.01188 0 0 0.001989975 -0.001989975
#>   [3,] 0.01176 0 0 0.002800000 -0.002800000
#>   [4,] 0.01164 0 0 0.003411744 -0.003411744
#>   [5,] 0.01152 0 0 0.003919184 -0.003919184
#>   [6,] 0.01140 0 0 0.004358899 -0.004358899
#>   [7,] 0.01128 0 0 0.004749737 -0.004749737
#>   [8,] 0.01116 0 0 0.005102940 -0.005102940
#>   [9,] 0.01104 0 0 0.005425864 -0.005425864
#>  [10,] 0.01092 0 0 0.005723635 -0.005723635
#>  [11,] 0.01080 0 0 0.006000000 -0.006000000
#>  [12,] 0.01068 0 0 0.006257795 -0.006257795
#>  [13,] 0.01056 0 0 0.006499231 -0.006499231
#>  [14,] 0.01044 0 0 0.006726069 -0.006726069
#>  [15,] 0.01032 0 0 0.006939741 -0.006939741
#>  [16,] 0.01020 0 0 0.007141428 -0.007141428
#>  [17,] 0.01008 0 0 0.007332121 -0.007332121
#>  [18,] 0.00996 0 0 0.007512656 -0.007512656
#>  [19,] 0.00984 0 0 0.007683749 -0.007683749
#>  [20,] 0.00972 0 0 0.007846018 -0.007846018
#>  [21,] 0.00960 0 0 0.008000000 -0.008000000
#>  [22,] 0.00948 0 0 0.008146165 -0.008146165
#>  [23,] 0.00936 0 0 0.008284926 -0.008284926
#>  [24,] 0.00924 0 0 0.008416650 -0.008416650
#>  [25,] 0.00912 0 0 0.008541663 -0.008541663
#>  [26,] 0.00900 0 0 0.008660254 -0.008660254
#>  [27,] 0.00888 0 0 0.008772685 -0.008772685
#>  [28,] 0.00876 0 0 0.008879189 -0.008879189
#>  [29,] 0.00864 0 0 0.008979978 -0.008979978
#>  [30,] 0.00852 0 0 0.009075241 -0.009075241
#>  [31,] 0.00840 0 0 0.009165151 -0.009165151
#>  [32,] 0.00828 0 0 0.009249865 -0.009249865
#>  [33,] 0.00816 0 0 0.009329523 -0.009329523
#>  [34,] 0.00804 0 0 0.009404254 -0.009404254
#>  [35,] 0.00792 0 0 0.009474175 -0.009474175
#>  [36,] 0.00780 0 0 0.009539392 -0.009539392
#>  [37,] 0.00768 0 0 0.009600000 -0.009600000
#>  [38,] 0.00756 0 0 0.009656086 -0.009656086
#>  [39,] 0.00744 0 0 0.009707729 -0.009707729
#>  [40,] 0.00732 0 0 0.009754999 -0.009754999
#>  [41,] 0.00720 0 0 0.009797959 -0.009797959
#>  [42,] 0.00708 0 0 0.009836666 -0.009836666
#>  [43,] 0.00696 0 0 0.009871170 -0.009871170
#>  [44,] 0.00684 0 0 0.009901515 -0.009901515
#>  [45,] 0.00672 0 0 0.009927739 -0.009927739
#>  [46,] 0.00660 0 0 0.009949874 -0.009949874
#>  [47,] 0.00648 0 0 0.009967949 -0.009967949
#>  [48,] 0.00636 0 0 0.009981984 -0.009981984
#>  [49,] 0.00624 0 0 0.009991997 -0.009991997
#>  [50,] 0.00612 0 0 0.009998000 -0.009998000
#>  [51,] 0.00600 0 0 0.010000000 -0.010000000
#>  [52,] 0.00588 0 0 0.009998000 -0.009998000
#>  [53,] 0.00576 0 0 0.009991997 -0.009991997
#>  [54,] 0.00564 0 0 0.009981984 -0.009981984
#>  [55,] 0.00552 0 0 0.009967949 -0.009967949
#>  [56,] 0.00540 0 0 0.009949874 -0.009949874
#>  [57,] 0.00528 0 0 0.009927739 -0.009927739
#>  [58,] 0.00516 0 0 0.009901515 -0.009901515
#>  [59,] 0.00504 0 0 0.009871170 -0.009871170
#>  [60,] 0.00492 0 0 0.009836666 -0.009836666
#>  [61,] 0.00480 0 0 0.009797959 -0.009797959
#>  [62,] 0.00468 0 0 0.009754999 -0.009754999
#>  [63,] 0.00456 0 0 0.009707729 -0.009707729
#>  [64,] 0.00444 0 0 0.009656086 -0.009656086
#>  [65,] 0.00432 0 0 0.009600000 -0.009600000
#>  [66,] 0.00420 0 0 0.009539392 -0.009539392
#>  [67,] 0.00408 0 0 0.009474175 -0.009474175
#>  [68,] 0.00396 0 0 0.009404254 -0.009404254
#>  [69,] 0.00384 0 0 0.009329523 -0.009329523
#>  [70,] 0.00372 0 0 0.009249865 -0.009249865
#>  [71,] 0.00360 0 0 0.009165151 -0.009165151
#>  [72,] 0.00348 0 0 0.009075241 -0.009075241
#>  [73,] 0.00336 0 0 0.008979978 -0.008979978
#>  [74,] 0.00324 0 0 0.008879189 -0.008879189
#>  [75,] 0.00312 0 0 0.008772685 -0.008772685
#>  [76,] 0.00300 0 0 0.008660254 -0.008660254
#>  [77,] 0.00288 0 0 0.008541663 -0.008541663
#>  [78,] 0.00276 0 0 0.008416650 -0.008416650
#>  [79,] 0.00264 0 0 0.008284926 -0.008284926
#>  [80,] 0.00252 0 0 0.008146165 -0.008146165
#>  [81,] 0.00240 0 0 0.008000000 -0.008000000
#>  [82,] 0.00228 0 0 0.007846018 -0.007846018
#>  [83,] 0.00216 0 0 0.007683749 -0.007683749
#>  [84,] 0.00204 0 0 0.007512656 -0.007512656
#>  [85,] 0.00192 0 0 0.007332121 -0.007332121
#>  [86,] 0.00180 0 0 0.007141428 -0.007141428
#>  [87,] 0.00168 0 0 0.006939741 -0.006939741
#>  [88,] 0.00156 0 0 0.006726069 -0.006726069
#>  [89,] 0.00144 0 0 0.006499231 -0.006499231
#>  [90,] 0.00132 0 0 0.006257795 -0.006257795
#>  [91,] 0.00120 0 0 0.006000000 -0.006000000
#>  [92,] 0.00108 0 0 0.005723635 -0.005723635
#>  [93,] 0.00096 0 0 0.005425864 -0.005425864
#>  [94,] 0.00084 0 0 0.005102940 -0.005102940
#>  [95,] 0.00072 0 0 0.004749737 -0.004749737
#>  [96,] 0.00060 0 0 0.004358899 -0.004358899
#>  [97,] 0.00048 0 0 0.003919184 -0.003919184
#>  [98,] 0.00036 0 0 0.003411744 -0.003411744
#>  [99,] 0.00024 0 0 0.002800000 -0.002800000
#> [100,] 0.00012 0 0 0.001989975 -0.001989975
#> [101,] 0.00000 0 0 0.000000000  0.000000000
#> 
#> Slot "shape_parameters":
#> $length
#> [1] 0.012
#> 
#> $radius
#>   [1] 0.000000000 0.001989975 0.002800000 0.003411744 0.003919184 0.004358899
#>   [7] 0.004749737 0.005102940 0.005425864 0.005723635 0.006000000 0.006257795
#>  [13] 0.006499231 0.006726069 0.006939741 0.007141428 0.007332121 0.007512656
#>  [19] 0.007683749 0.007846018 0.008000000 0.008146165 0.008284926 0.008416650
#>  [25] 0.008541663 0.008660254 0.008772685 0.008879189 0.008979978 0.009075241
#>  [31] 0.009165151 0.009249865 0.009329523 0.009404254 0.009474175 0.009539392
#>  [37] 0.009600000 0.009656086 0.009707729 0.009754999 0.009797959 0.009836666
#>  [43] 0.009871170 0.009901515 0.009927739 0.009949874 0.009967949 0.009981984
#>  [49] 0.009991997 0.009998000 0.010000000 0.009998000 0.009991997 0.009981984
#>  [55] 0.009967949 0.009949874 0.009927739 0.009901515 0.009871170 0.009836666
#>  [61] 0.009797959 0.009754999 0.009707729 0.009656086 0.009600000 0.009539392
#>  [67] 0.009474175 0.009404254 0.009329523 0.009249865 0.009165151 0.009075241
#>  [73] 0.008979978 0.008879189 0.008772685 0.008660254 0.008541663 0.008416650
#>  [79] 0.008284926 0.008146165 0.008000000 0.007846018 0.007683749 0.007512656
#>  [85] 0.007332121 0.007141428 0.006939741 0.006726069 0.006499231 0.006257795
#>  [91] 0.006000000 0.005723635 0.005425864 0.005102940 0.004749737 0.004358899
#>  [97] 0.003919184 0.003411744 0.002800000 0.001989975 0.000000000
#> 
#> $semimajor_length
#> [1] 0.01
#> 
#> $semiminor_length
#> [1] 0.006
#> 
#> $length_radius_ratio
#> [1] 1.2
#> 
#> $n_segments
#> [1] 100
#> 
#> $length_units
#> [1] "m"
#> 
#>