Creates a prolate spheroid.

Creates a prolate spheroid.

Usage

prolate_spheroid(
  length_body = NULL,
  radius_body = NULL,
  length_radius_ratio = NULL,
  semimajor_length = NULL,
  semiminor_length = NULL,
  n_segments = 100,
  length_units = "m"
)

Arguments

length_body

Semi-major axis length (m).

radius_body

Semi-minor axis length (m). This can also be stylized as the "maximum radius" of the scattering object.

length_radius_ratio

Optional ratio input when radius is not explicitly known.

semimajor_length

Optional alias for semi-major axis length (m).

semiminor_length

Optional alias for semi-minor axis length (m).

n_segments

Number of segments to discretize object shape. Defaults to 18 segments.

length_units

Units for body matrix (defaults to m).

Value

Creates the position vector for a prolate spheroid object of defined semi-major and -minor axes.

See also

Examples

prolate_spheroid(
  length_body = 0.04, radius_body = 0.004, n_segments = 60
)
#> An object of class "ProlateSpheroid"
#> Slot "position_matrix":
#>                  x y z          zU           zL
#>  [1,] 0.0400000000 0 0 0.000000000  0.000000000
#>  [2,] 0.0393333333 0 0 0.001024153 -0.001024153
#>  [3,] 0.0386666667 0 0 0.001436044 -0.001436044
#>  [4,] 0.0380000000 0 0 0.001743560 -0.001743560
#>  [5,] 0.0373333333 0 0 0.001995551 -0.001995551
#>  [6,] 0.0366666667 0 0 0.002211083 -0.002211083
#>  [7,] 0.0360000000 0 0 0.002400000 -0.002400000
#>  [8,] 0.0353333333 0 0 0.002568181 -0.002568181
#>  [9,] 0.0346666667 0 0 0.002719477 -0.002719477
#> [10,] 0.0340000000 0 0 0.002856571 -0.002856571
#> [11,] 0.0333333333 0 0 0.002981424 -0.002981424
#> [12,] 0.0326666667 0 0 0.003095516 -0.003095516
#> [13,] 0.0320000000 0 0 0.003200000 -0.003200000
#> [14,] 0.0313333333 0 0 0.003295789 -0.003295789
#> [15,] 0.0306666667 0 0 0.003383621 -0.003383621
#> [16,] 0.0300000000 0 0 0.003464102 -0.003464102
#> [17,] 0.0293333333 0 0 0.003537733 -0.003537733
#> [18,] 0.0286666667 0 0 0.003604935 -0.003604935
#> [19,] 0.0280000000 0 0 0.003666061 -0.003666061
#> [20,] 0.0273333333 0 0 0.003721410 -0.003721410
#> [21,] 0.0266666667 0 0 0.003771236 -0.003771236
#> [22,] 0.0260000000 0 0 0.003815757 -0.003815757
#> [23,] 0.0253333333 0 0 0.003855155 -0.003855155
#> [24,] 0.0246666667 0 0 0.003889587 -0.003889587
#> [25,] 0.0240000000 0 0 0.003919184 -0.003919184
#> [26,] 0.0233333333 0 0 0.003944053 -0.003944053
#> [27,] 0.0226666667 0 0 0.003964285 -0.003964285
#> [28,] 0.0220000000 0 0 0.003979950 -0.003979950
#> [29,] 0.0213333333 0 0 0.003991101 -0.003991101
#> [30,] 0.0206666667 0 0 0.003997777 -0.003997777
#> [31,] 0.0200000000 0 0 0.004000000 -0.004000000
#> [32,] 0.0193333333 0 0 0.003997777 -0.003997777
#> [33,] 0.0186666667 0 0 0.003991101 -0.003991101
#> [34,] 0.0180000000 0 0 0.003979950 -0.003979950
#> [35,] 0.0173333333 0 0 0.003964285 -0.003964285
#> [36,] 0.0166666667 0 0 0.003944053 -0.003944053
#> [37,] 0.0160000000 0 0 0.003919184 -0.003919184
#> [38,] 0.0153333333 0 0 0.003889587 -0.003889587
#> [39,] 0.0146666667 0 0 0.003855155 -0.003855155
#> [40,] 0.0140000000 0 0 0.003815757 -0.003815757
#> [41,] 0.0133333333 0 0 0.003771236 -0.003771236
#> [42,] 0.0126666667 0 0 0.003721410 -0.003721410
#> [43,] 0.0120000000 0 0 0.003666061 -0.003666061
#> [44,] 0.0113333333 0 0 0.003604935 -0.003604935
#> [45,] 0.0106666667 0 0 0.003537733 -0.003537733
#> [46,] 0.0100000000 0 0 0.003464102 -0.003464102
#> [47,] 0.0093333333 0 0 0.003383621 -0.003383621
#> [48,] 0.0086666667 0 0 0.003295789 -0.003295789
#> [49,] 0.0080000000 0 0 0.003200000 -0.003200000
#> [50,] 0.0073333333 0 0 0.003095516 -0.003095516
#> [51,] 0.0066666667 0 0 0.002981424 -0.002981424
#> [52,] 0.0060000000 0 0 0.002856571 -0.002856571
#> [53,] 0.0053333333 0 0 0.002719477 -0.002719477
#> [54,] 0.0046666667 0 0 0.002568181 -0.002568181
#> [55,] 0.0040000000 0 0 0.002400000 -0.002400000
#> [56,] 0.0033333333 0 0 0.002211083 -0.002211083
#> [57,] 0.0026666667 0 0 0.001995551 -0.001995551
#> [58,] 0.0020000000 0 0 0.001743560 -0.001743560
#> [59,] 0.0013333333 0 0 0.001436044 -0.001436044
#> [60,] 0.0006666667 0 0 0.001024153 -0.001024153
#> [61,] 0.0000000000 0 0 0.000000000  0.000000000
#> 
#> Slot "shape_parameters":
#> $length
#> [1] 0.04
#> 
#> $radius
#>  [1] 0.000000000 0.001024153 0.001436044 0.001743560 0.001995551 0.002211083
#>  [7] 0.002400000 0.002568181 0.002719477 0.002856571 0.002981424 0.003095516
#> [13] 0.003200000 0.003295789 0.003383621 0.003464102 0.003537733 0.003604935
#> [19] 0.003666061 0.003721410 0.003771236 0.003815757 0.003855155 0.003889587
#> [25] 0.003919184 0.003944053 0.003964285 0.003979950 0.003991101 0.003997777
#> [31] 0.004000000 0.003997777 0.003991101 0.003979950 0.003964285 0.003944053
#> [37] 0.003919184 0.003889587 0.003855155 0.003815757 0.003771236 0.003721410
#> [43] 0.003666061 0.003604935 0.003537733 0.003464102 0.003383621 0.003295789
#> [49] 0.003200000 0.003095516 0.002981424 0.002856571 0.002719477 0.002568181
#> [55] 0.002400000 0.002211083 0.001995551 0.001743560 0.001436044 0.001024153
#> [61] 0.000000000
#> 
#> $semimajor_length
#> [1] 0.02
#> 
#> $semiminor_length
#> [1] 0.004
#> 
#> $length_radius_ratio
#> [1] 10
#> 
#> $n_segments
#> [1] 60
#> 
#> $length_units
#> [1] "m"
#> 
#> 
prolate_spheroid(
  semimajor_length = 0.05, semiminor_length = 0.003
)
#> An object of class "ProlateSpheroid"
#> Slot "position_matrix":
#>            x y z           zU            zL
#>   [1,] 0.100 0 0 0.0000000000  0.0000000000
#>   [2,] 0.099 0 0 0.0005969925 -0.0005969925
#>   [3,] 0.098 0 0 0.0008400000 -0.0008400000
#>   [4,] 0.097 0 0 0.0010235233 -0.0010235233
#>   [5,] 0.096 0 0 0.0011757551 -0.0011757551
#>   [6,] 0.095 0 0 0.0013076697 -0.0013076697
#>   [7,] 0.094 0 0 0.0014249211 -0.0014249211
#>   [8,] 0.093 0 0 0.0015308821 -0.0015308821
#>   [9,] 0.092 0 0 0.0016277592 -0.0016277592
#>  [10,] 0.091 0 0 0.0017170906 -0.0017170906
#>  [11,] 0.090 0 0 0.0018000000 -0.0018000000
#>  [12,] 0.089 0 0 0.0018773385 -0.0018773385
#>  [13,] 0.088 0 0 0.0019497692 -0.0019497692
#>  [14,] 0.087 0 0 0.0020178206 -0.0020178206
#>  [15,] 0.086 0 0 0.0020819222 -0.0020819222
#>  [16,] 0.085 0 0 0.0021424285 -0.0021424285
#>  [17,] 0.084 0 0 0.0021996363 -0.0021996363
#>  [18,] 0.083 0 0 0.0022537968 -0.0022537968
#>  [19,] 0.082 0 0 0.0023051247 -0.0023051247
#>  [20,] 0.081 0 0 0.0023538054 -0.0023538054
#>  [21,] 0.080 0 0 0.0024000000 -0.0024000000
#>  [22,] 0.079 0 0 0.0024438494 -0.0024438494
#>  [23,] 0.078 0 0 0.0024854778 -0.0024854778
#>  [24,] 0.077 0 0 0.0025249950 -0.0025249950
#>  [25,] 0.076 0 0 0.0025624988 -0.0025624988
#>  [26,] 0.075 0 0 0.0025980762 -0.0025980762
#>  [27,] 0.074 0 0 0.0026318055 -0.0026318055
#>  [28,] 0.073 0 0 0.0026637567 -0.0026637567
#>  [29,] 0.072 0 0 0.0026939933 -0.0026939933
#>  [30,] 0.071 0 0 0.0027225723 -0.0027225723
#>  [31,] 0.070 0 0 0.0027495454 -0.0027495454
#>  [32,] 0.069 0 0 0.0027749595 -0.0027749595
#>  [33,] 0.068 0 0 0.0027988569 -0.0027988569
#>  [34,] 0.067 0 0 0.0028212763 -0.0028212763
#>  [35,] 0.066 0 0 0.0028422526 -0.0028422526
#>  [36,] 0.065 0 0 0.0028618176 -0.0028618176
#>  [37,] 0.064 0 0 0.0028800000 -0.0028800000
#>  [38,] 0.063 0 0 0.0028968258 -0.0028968258
#>  [39,] 0.062 0 0 0.0029123187 -0.0029123187
#>  [40,] 0.061 0 0 0.0029264996 -0.0029264996
#>  [41,] 0.060 0 0 0.0029393877 -0.0029393877
#>  [42,] 0.059 0 0 0.0029509998 -0.0029509998
#>  [43,] 0.058 0 0 0.0029613510 -0.0029613510
#>  [44,] 0.057 0 0 0.0029704545 -0.0029704545
#>  [45,] 0.056 0 0 0.0029783217 -0.0029783217
#>  [46,] 0.055 0 0 0.0029849623 -0.0029849623
#>  [47,] 0.054 0 0 0.0029903846 -0.0029903846
#>  [48,] 0.053 0 0 0.0029945951 -0.0029945951
#>  [49,] 0.052 0 0 0.0029975990 -0.0029975990
#>  [50,] 0.051 0 0 0.0029993999 -0.0029993999
#>  [51,] 0.050 0 0 0.0030000000 -0.0030000000
#>  [52,] 0.049 0 0 0.0029993999 -0.0029993999
#>  [53,] 0.048 0 0 0.0029975990 -0.0029975990
#>  [54,] 0.047 0 0 0.0029945951 -0.0029945951
#>  [55,] 0.046 0 0 0.0029903846 -0.0029903846
#>  [56,] 0.045 0 0 0.0029849623 -0.0029849623
#>  [57,] 0.044 0 0 0.0029783217 -0.0029783217
#>  [58,] 0.043 0 0 0.0029704545 -0.0029704545
#>  [59,] 0.042 0 0 0.0029613510 -0.0029613510
#>  [60,] 0.041 0 0 0.0029509998 -0.0029509998
#>  [61,] 0.040 0 0 0.0029393877 -0.0029393877
#>  [62,] 0.039 0 0 0.0029264996 -0.0029264996
#>  [63,] 0.038 0 0 0.0029123187 -0.0029123187
#>  [64,] 0.037 0 0 0.0028968258 -0.0028968258
#>  [65,] 0.036 0 0 0.0028800000 -0.0028800000
#>  [66,] 0.035 0 0 0.0028618176 -0.0028618176
#>  [67,] 0.034 0 0 0.0028422526 -0.0028422526
#>  [68,] 0.033 0 0 0.0028212763 -0.0028212763
#>  [69,] 0.032 0 0 0.0027988569 -0.0027988569
#>  [70,] 0.031 0 0 0.0027749595 -0.0027749595
#>  [71,] 0.030 0 0 0.0027495454 -0.0027495454
#>  [72,] 0.029 0 0 0.0027225723 -0.0027225723
#>  [73,] 0.028 0 0 0.0026939933 -0.0026939933
#>  [74,] 0.027 0 0 0.0026637567 -0.0026637567
#>  [75,] 0.026 0 0 0.0026318055 -0.0026318055
#>  [76,] 0.025 0 0 0.0025980762 -0.0025980762
#>  [77,] 0.024 0 0 0.0025624988 -0.0025624988
#>  [78,] 0.023 0 0 0.0025249950 -0.0025249950
#>  [79,] 0.022 0 0 0.0024854778 -0.0024854778
#>  [80,] 0.021 0 0 0.0024438494 -0.0024438494
#>  [81,] 0.020 0 0 0.0024000000 -0.0024000000
#>  [82,] 0.019 0 0 0.0023538054 -0.0023538054
#>  [83,] 0.018 0 0 0.0023051247 -0.0023051247
#>  [84,] 0.017 0 0 0.0022537968 -0.0022537968
#>  [85,] 0.016 0 0 0.0021996363 -0.0021996363
#>  [86,] 0.015 0 0 0.0021424285 -0.0021424285
#>  [87,] 0.014 0 0 0.0020819222 -0.0020819222
#>  [88,] 0.013 0 0 0.0020178206 -0.0020178206
#>  [89,] 0.012 0 0 0.0019497692 -0.0019497692
#>  [90,] 0.011 0 0 0.0018773385 -0.0018773385
#>  [91,] 0.010 0 0 0.0018000000 -0.0018000000
#>  [92,] 0.009 0 0 0.0017170906 -0.0017170906
#>  [93,] 0.008 0 0 0.0016277592 -0.0016277592
#>  [94,] 0.007 0 0 0.0015308821 -0.0015308821
#>  [95,] 0.006 0 0 0.0014249211 -0.0014249211
#>  [96,] 0.005 0 0 0.0013076697 -0.0013076697
#>  [97,] 0.004 0 0 0.0011757551 -0.0011757551
#>  [98,] 0.003 0 0 0.0010235233 -0.0010235233
#>  [99,] 0.002 0 0 0.0008400000 -0.0008400000
#> [100,] 0.001 0 0 0.0005969925 -0.0005969925
#> [101,] 0.000 0 0 0.0000000000  0.0000000000
#> 
#> Slot "shape_parameters":
#> $length
#> [1] 0.1
#> 
#> $radius
#>   [1] 0.0000000000 0.0005969925 0.0008400000 0.0010235233 0.0011757551
#>   [6] 0.0013076697 0.0014249211 0.0015308821 0.0016277592 0.0017170906
#>  [11] 0.0018000000 0.0018773385 0.0019497692 0.0020178206 0.0020819222
#>  [16] 0.0021424285 0.0021996363 0.0022537968 0.0023051247 0.0023538054
#>  [21] 0.0024000000 0.0024438494 0.0024854778 0.0025249950 0.0025624988
#>  [26] 0.0025980762 0.0026318055 0.0026637567 0.0026939933 0.0027225723
#>  [31] 0.0027495454 0.0027749595 0.0027988569 0.0028212763 0.0028422526
#>  [36] 0.0028618176 0.0028800000 0.0028968258 0.0029123187 0.0029264996
#>  [41] 0.0029393877 0.0029509998 0.0029613510 0.0029704545 0.0029783217
#>  [46] 0.0029849623 0.0029903846 0.0029945951 0.0029975990 0.0029993999
#>  [51] 0.0030000000 0.0029993999 0.0029975990 0.0029945951 0.0029903846
#>  [56] 0.0029849623 0.0029783217 0.0029704545 0.0029613510 0.0029509998
#>  [61] 0.0029393877 0.0029264996 0.0029123187 0.0028968258 0.0028800000
#>  [66] 0.0028618176 0.0028422526 0.0028212763 0.0027988569 0.0027749595
#>  [71] 0.0027495454 0.0027225723 0.0026939933 0.0026637567 0.0026318055
#>  [76] 0.0025980762 0.0025624988 0.0025249950 0.0024854778 0.0024438494
#>  [81] 0.0024000000 0.0023538054 0.0023051247 0.0022537968 0.0021996363
#>  [86] 0.0021424285 0.0020819222 0.0020178206 0.0019497692 0.0018773385
#>  [91] 0.0018000000 0.0017170906 0.0016277592 0.0015308821 0.0014249211
#>  [96] 0.0013076697 0.0011757551 0.0010235233 0.0008400000 0.0005969925
#> [101] 0.0000000000
#> 
#> $semimajor_length
#> [1] 0.05
#> 
#> $semiminor_length
#> [1] 0.003
#> 
#> $length_radius_ratio
#> [1] 33.33333
#> 
#> $n_segments
#> [1] 100
#> 
#> $length_units
#> [1] "m"
#> 
#>