A wrapper function that automatically creates generalized and/or canonical shapes for TS modeling.

A wrapper function that automatically creates generalized and/or canonical shapes for TS modeling.

Usage

create_shape(shape, ...)

Arguments

shape

Shape. Details for shape specification are provided under 'Details', including mandatory arguments.

...

Additional input arguments for subsequent shape generation functions.

Value

Chu, D., Foote, K.G., and Stanton, T.K. 1993. Further analysis of target strength measurements of Antarctic krill at 38 and 120 kHz: Comparison and deformed cylinder model and inference of orientation distribution. The Journal of the Acoustical Society of America, 93(5): 2985-2988. https://doi.org/10.1121/1.405818

Smith, J.N., Ressler, P.H., and Warren, J.D. 2013. A distorted wave Born approximation target strength model for Bering Sea euphausiids. ICES Journal of Marine Science, 70(1): 204-214. https://doi.org/10.1093/icesjms/fss140

Details

The shape argument specifies what shape for the function to generate into the desired shape for TS modeling. Options currently include:

Object shape	shape = ...		Parameters	Root function

Discrete/tapered cylinder	"cylinder"		length, radius	cylinder(...)

Polynomial cylinder	"polynomial_cylinder"		length, radius, polynomial	polynomial_cylinder(...)

Prolate spheroid	"prolate_speroid"		length, radius	prolate_spheroid(...)

Sphere	"sphere"		radius	sphere(...)

Model Parameter Definitions

length: the x-axis length of the shape.

radius: the radius of the shape when applicable.

length_radius_ratio: the length-to-radius ratio (L/A), which specifically refers to the radius at the mid-point of the cylinder and should be the maximum value. A typical L/A ratio in the literature is 16 for krill.

taper: the taper order (n), which parameterizes the tapering function reported by Chu et al. (1993) to create a tapered cylinder. The tapering order will converge on a prolate and oblate spheroid when L > 2a and L < 2a, respectively, and n = 2. A typical taper order in the literature is 10.

polynomial: the vector of arbitrary polynomial coefficients to generate a deformed cylinder as reported by Smith et al. (2013). Although listed as a mandatory argument for the polynomial cylinder function, it has a default setting that uses the sixth-degree polynomial coefficients reported by Smith et al. (2013).