Cylindrical Bessel function of the first kind and its respective derivatives

Calculate the cylindrical Bessel function of the first kind (jc) and both its first (jcd) and second (jcdd) derivatives.

Usage

jc(l, n)

jcd(l, n)

jcdd(l, n)

Arguments

l

An integer or fractional order

n

A complex or real argument

Value

Calculates the cylindrical Bessel function of the first kind (\(J_v\)) and its respective derivatives.

References

Amos D.E., "AMOS, A Portable Package for Bessel Functions of a Complex Argument and Nonnegative Order", http://netlib.org/amos